Relocated deployment of commands to index.js and remove deploy-commands.js file along with npm register script
