--- deploy-commands.js
@@ -1,19 +1,0 @@
-const { REST } = require('@discordjs/rest');
-const { Routes } = require('discord-api-types/v9');
-const dotenv = require('dotenv');
-dotenv.config();
-const fs = require('fs');
-
-const commands = [];
-const commandFiles = fs.readdirSync('./commands').filter(file => file.endsWith('.js'));
-
-for (const file of commandFiles) {
-  const command = require(`./commands/${file}`);
-  commands.push(command.data.toJSON());
-}
-
-const rest = new REST({ version: '9' }).setToken(process.env.DISCORD_TOKEN);
-
-rest.put(Routes.applicationGuildCommands(process.env.clientId, process.env.guildId), { body: commands })
-  .then(() => console.log('Successfully registered application commands.')) //eslint-disable-line
-  .catch(err => console.error(err));
