+++ package.json
@@ -5,8 +5,7 @@
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node .",
-    "register": "node deploy-commands.js"
+    "start": "node ."
   },
   "repository": {
     "type": "git",
